move classic admin to separate module implement backward compatibility
