+++ crates/fluvio-sc-schema/src/objects/mod.rs
@@ -2,6 +2,8 @@ mod create;
 mod delete;
 mod list;
 mod watch;
+mod metadata;
+mod classic;
 
 pub use create::*;
 pub use delete::*;
@@ -12,563 +14,5 @@ pub use metadata::*;
 pub(crate) const COMMON_VERSION: i16 = 11; // from now, we use a single version for all objects
 pub(crate) const DYN_OBJ: i16 = 11; // version indicate dynamic object
 
-mod metadata {
-
-    use std::convert::{TryFrom, TryInto};
-    use std::fmt::{Debug, Display};
-    use std::io::{Error as IoError, Cursor};
-    use std::io::ErrorKind;
-
-    use anyhow::Result;
-
-    use fluvio_protocol::{Encoder, Decoder, ByteBuf, Version};
-
-    use fluvio_controlplane_metadata::store::MetadataStoreObject;
-    use fluvio_controlplane_metadata::core::{MetadataContext, MetadataItem};
-
-    use crate::AdminSpec;
-    use crate::core::Spec;
-
-    use super::{COMMON_VERSION, DYN_OBJ};
-
-    #[derive(Encoder, Decoder, Default, Clone, Debug)]
-    #[cfg_attr(
-        feature = "use_serde",
-        derive(serde::Serialize, serde::Deserialize),
-        serde(rename_all = "camelCase")
-    )]
-    pub struct Metadata<S>
-    where
-        S: Spec + Encoder + Decoder,
-        S::Status: Encoder + Decoder,
-    {
-        pub name: String,
-        pub spec: S,
-        pub status: S::Status,
-    }
-
-    impl<S, C> From<MetadataStoreObject<S, C>> for Metadata<S>
-    where
-        S: Spec + Encoder + Decoder,
-        S::IndexKey: ToString,
-        S::Status: Encoder + Decoder,
-        C: MetadataItem,
-    {
-        fn from(meta: MetadataStoreObject<S, C>) -> Self {
-            Self {
-                name: meta.key.to_string(),
-                spec: meta.spec,
-                status: meta.status,
-            }
-        }
-    }
-
-    impl<S> Metadata<S>
-    where
-        S: AdminSpec + Encoder + Decoder,
-        S::Status: Encoder + Decoder,
-    {
-        pub fn summary(self) -> Self {
-            Self {
-                name: self.name,
-                spec: self.spec.summary(),
-                status: self.status,
-            }
-        }
-    }
-
-    impl<S, C> TryFrom<Metadata<S>> for MetadataStoreObject<S, C>
-    where
-        S: Spec + Encoder + Decoder,
-        S::Status: Encoder + Decoder,
-        C: MetadataItem,
-        <S as Spec>::IndexKey: TryFrom<String>,
-        <<S as Spec>::IndexKey as TryFrom<String>>::Error: Display,
-    {
-        type Error = IoError;
-
-        fn try_from(value: Metadata<S>) -> Result<Self, Self::Error> {
-            Ok(Self {
-                spec: value.spec,
-                status: value.status,
-                key: value.name.try_into().map_err(|err| {
-                    IoError::new(ErrorKind::InvalidData, format!("problem converting: {err}"))
-                })?,
-                ctx: MetadataContext::default(),
-            })
-        }
-    }
-
-    /// Type encoded buffer, it uses type label to determine type
-    #[derive(Debug, Default)]
-    pub struct TypeBuffer {
-        ty: String,
-        buf: ByteBuf,
-    }
-
-    impl TypeBuffer {
-        // encode admin spec into a request
-        pub fn encode<S, I>(input: I, version: Version) -> Result<Self>
-        where
-            S: Spec,
-            I: Encoder,
-        {
-            let ty = S::LABEL.to_owned();
-            let mut buf = vec![];
-            input.encode(&mut buf, version)?;
-            Ok(Self {
-                ty,
-                buf: ByteBuf::from(buf),
-            })
-        }
-
-        // check if this object is kind of spec
-        pub fn is_kind_of<S: Spec>(&self) -> bool {
-            self.ty == S::LABEL
-        }
-
-        // downcast to specific spec type and return object
-        // if doens't match to ty, return None
-        pub fn downcast<S, O>(&self) -> Result<Option<O>>
-        where
-            S: Spec,
-            O: Decoder + Debug,
-        {
-            if self.is_kind_of::<S>() {
-                println!("is kind of: {:#?}", S::LABEL);
-                let mut buf = Cursor::new(self.buf.as_ref());
-                Ok(Some(O::decode_from(&mut buf, COMMON_VERSION)?))
-            } else {
-                println!("not kind of: {:#?}", S::LABEL);
-                Ok(None)
-            }
-        }
-    }
-
-    impl Encoder for TypeBuffer {
-        fn write_size(&self, version: Version) -> usize {
-            self.ty.write_size(version)
-                + self.buf.len()
-                + (if version >= DYN_OBJ {
-                    let u32 = 0;
-                    u32.write_size(version)
-                } else {
-                    0
-                })
-        }
-
-        fn encode<T>(&self, dest: &mut T, version: Version) -> std::result::Result<(), IoError>
-        where
-            T: fluvio_protocol::bytes::BufMut,
-        {
-            self.ty.encode(dest, version)?;
-            if version >= DYN_OBJ {
-                let len: u32 = self.buf.len() as u32;
-                len.encode(dest, version)?; // write len
-                println!("encoded len: {:#?}", len);
-            }
-            dest.put(self.buf.as_ref());
-            Ok(())
-        }
-    }
-
-    impl Decoder for TypeBuffer {
-        fn decode<T>(&mut self, src: &mut T, version: Version) -> std::result::Result<(), IoError>
-        where
-            T: fluvio_protocol::bytes::Buf,
-        {
-            self.ty.decode(src, version)?;
-            tracing::trace!(ty = self.ty, "decoded type");
-            println!("decoded type: {:#?}", self.ty);
-
-            if version >= DYN_OBJ {
-                let mut len: u32 = 0;
-                len.decode(src, version)?;
-                tracing::trace!(len, "decoded len");
-                println!("copy bytes: {:#?}", len);
-                self.buf = src.copy_to_bytes(len as usize).into();
-            } else {
-                // old version, we need to compute len using old way which may not be reliable
-                self.buf = src.copy_to_bytes(src.remaining()).into();
-            }
-
-            Ok(())
-        }
-    }
-}
-
 #[cfg(test)]
-mod object_macro {
-    use std::io::Cursor;
-
-    use fluvio_controlplane_metadata::topic::TopicSpec;
-    use fluvio_protocol::{Encoder, Decoder};
-
-    use crate::{
-        objects::{ListRequest, ObjectApiListRequest, COMMON_VERSION},
-        TryEncodableFrom,
-    };
-
-    /// carry from prev version for compatibility test
-    macro_rules! ObjectApiEnum {
-        ($api:ident) => {
-
-            paste::paste! {
-
-
-                #[derive(Debug)]
-                pub enum [<ObjectApiOld $api>] {
-                    Topic($api<crate::topic::TopicSpec>),
-                    Spu($api<crate::spu::SpuSpec>)
-                }
-
-                impl Default for [<ObjectApiOld $api>] {
-                    fn default() -> Self {
-                        Self::Topic($api::<crate::topic::TopicSpec>::default())
-                    }
-                }
-
-                impl [<ObjectApiOld $api>] {
-                    fn type_string(&self) -> &'static str {
-                        use fluvio_controlplane_metadata::core::Spec;
-                        match self {
-                            Self::Topic(_) => crate::topic::TopicSpec::LABEL,
-                            Self::Spu(_) => crate::spu::SpuSpec::LABEL,
-                        }
-                    }
-                }
-
-                impl  fluvio_protocol::Encoder for [<ObjectApiOld $api>] {
-
-                    fn write_size(&self, version: fluvio_protocol::Version) -> usize {
-                        let type_size = self.type_string().to_owned().write_size(version);
-
-                        type_size
-                            + match self {
-                                Self::Topic(s) => s.write_size(version),
-                                Self::Spu(s) => s.write_size(version),
-                            }
-                    }
-
-                    fn encode<T>(&self, dest: &mut T, version: fluvio_protocol::Version) -> Result<(), std::io::Error>
-                    where
-                        T: fluvio_protocol::bytes::BufMut,
-                    {
-                        let ty = self.type_string().to_owned();
-
-                        tracing::trace!(%ty,len = self.write_size(version),"encoding objects");
-                        ty.encode(dest, version)?;
-
-                        match self {
-                            Self::Topic(s) => s.encode(dest, version)?,
-                            Self::Spu(s) => s.encode(dest, version)?,
-
-                        }
-
-                        Ok(())
-                    }
-
-                }
-
-
-                impl  fluvio_protocol::Decoder for [<ObjectApiOld $api>] {
-
-                    fn decode<T>(&mut self, src: &mut T, version: fluvio_protocol::Version) -> Result<(),std::io::Error>
-                    where
-                        T: fluvio_protocol::bytes::Buf
-                    {
-                        use fluvio_controlplane_metadata::core::Spec;
-
-                        let mut typ = "".to_owned();
-                        typ.decode(src, version)?;
-                        tracing::trace!(%typ,"decoded type");
-
-                        match typ.as_ref() {
-                            crate::topic::TopicSpec::LABEL => {
-                                tracing::trace!("detected topic");
-                                let mut request = $api::<crate::topic::TopicSpec>::default();
-                                request.decode(src, version)?;
-                                *self = Self::Topic(request);
-                                return Ok(())
-                            }
-
-                            crate::spu::SpuSpec::LABEL  => {
-                                tracing::trace!("detected spu");
-                                let mut request = $api::<crate::spu::SpuSpec>::default();
-                                request.decode(src, version)?;
-                                *self = Self::Spu(request);
-                                return Ok(())
-                            }
-
-
-                            // Unexpected type
-                            _ => Err(std::io::Error::new(
-                                std::io::ErrorKind::InvalidData,
-                                format!("invalid object type {:#?}", typ),
-                            ))
-                        }
-                    }
-
-                }
-            }
-        }
-    }
-
-    ObjectApiEnum!(ListRequest);
-
-    /// encode same request using new and old version
-    #[test]
-    fn test_req_upcast_encoding() {
-        let raw_req: ListRequest<TopicSpec> = ListRequest::new("test", false);
-        // upcast
-        let list_request =
-            ObjectApiListRequest::try_encode_from(raw_req, COMMON_VERSION - 1).expect("encoded");
-        let mut new_dest = vec![];
-        list_request
-            .encode(&mut new_dest, COMMON_VERSION - 1)
-            .expect("encoding");
-
-        let raw_req2: ListRequest<TopicSpec> = ListRequest::new("test", false);
-        let old_topic_request = ObjectApiOldListRequest::Topic(raw_req2);
-        let mut old_dest: Vec<u8> = vec![];
-        old_topic_request
-            .encode(&mut old_dest, COMMON_VERSION - 1)
-            .expect("encoding");
-
-        //  assert_eq!(new_dest.len(),20);
-        assert_eq!(old_dest, new_dest);
-    }
-
-    // encoding and decoding request
-    #[test]
-    fn test_req_encoding_decoding() {
-        let raw_req: ListRequest<TopicSpec> = ListRequest::new("test", false);
-
-        let test_request =
-            ObjectApiListRequest::try_encode_from(raw_req, COMMON_VERSION).expect("encoded");
-        let mut dest = vec![];
-        test_request
-            .encode(&mut dest, COMMON_VERSION)
-            .expect("encoding");
-
-        let recovered_request =
-            ObjectApiListRequest::decode_from(&mut Cursor::new(dest), COMMON_VERSION)
-                .expect("decode");
-
-        let downcast =
-            recovered_request.downcast().expect("downcast") as Option<ListRequest<TopicSpec>>;
-        assert!(downcast.is_some());
-    }
-
-    #[test]
-    fn test_req_old_to_new() {
-        let raw_req: ListRequest<TopicSpec> = ListRequest::new(vec![], false);
-
-        let old_topic_request = ObjectApiOldListRequest::Topic(raw_req);
-        let mut dest = vec![];
-        old_topic_request
-            .encode(&mut dest, COMMON_VERSION)
-            .expect("encoding");
-
-        let new_topic_request =
-            ObjectApiListRequest::decode_from(&mut Cursor::new(dest), COMMON_VERSION - 1)
-                .expect("decode");
-
-        let downcast =
-            new_topic_request.downcast().expect("downcast") as Option<ListRequest<TopicSpec>>;
-        assert!(downcast.is_some());
-    }
-}
-
-#[cfg(test)]
-mod test {
-
-    use std::io::Cursor;
-
-    use fluvio_protocol::api::{RequestHeader, ResponseMessage, RequestMessage};
-    use fluvio_protocol::{Encoder, Decoder};
-    use fluvio_protocol::api::Request;
-    use fluvio_controlplane_metadata::spu::SpuStatus;
-
-    use crate::TryEncodableFrom;
-    use crate::objects::{
-        Metadata, MetadataUpdate, ListResponse, ObjectApiWatchRequest, ObjectApiListResponse,
-    };
-
-    use crate::topic::TopicSpec;
-    use crate::customspu::CustomSpuSpec;
-
-    use super::{
-        ListRequest, ObjectApiListRequest, WatchResponse, ObjectApiWatchResponse, COMMON_VERSION,
-    };
-
-    fn create_req() -> ObjectApiListRequest {
-        let list_request: ListRequest<TopicSpec> = ListRequest::new(vec![], false);
-        ObjectApiListRequest::try_encode_from(list_request, COMMON_VERSION).expect("encode")
-    }
-
-    fn create_res() -> ObjectApiWatchResponse {
-        let update = MetadataUpdate {
-            epoch: 2,
-            changes: vec![],
-            all: vec![],
-        };
-        let watch_response: WatchResponse<TopicSpec> = WatchResponse::new(update);
-        ObjectApiWatchResponse::try_encode_from(watch_response, COMMON_VERSION).expect("encode")
-    }
-
-    #[test]
-    fn test_from() {
-        let req = create_req();
-        assert!((req.downcast().expect("downcast") as Option<ListRequest<TopicSpec>>).is_some());
-    }
-
-    #[test]
-    fn test_encode_decoding_dynamic() {
-        use fluvio_protocol::api::Request;
-
-        let req = create_req();
-
-        let mut req_msg = RequestMessage::new_request(req);
-        req_msg
-            .get_mut_header()
-            .set_client_id("test")
-            .set_api_version(ObjectApiListRequest::API_KEY as i16);
-
-        let mut src = vec![];
-        req_msg.encode(&mut src, 0).expect("encoding");
-
-        let dec_msg: RequestMessage<ObjectApiListRequest> = RequestMessage::decode_from(
-            &mut Cursor::new(&src),
-            ObjectApiListRequest::API_KEY as i16,
-        )
-        .expect("decode");
-        assert!(
-            (dec_msg.request.downcast().expect("downcast") as Option<ListRequest<TopicSpec>>)
-                .is_some()
-        );
-    }
-
-    // test encoding and decoding of metadata update
-    #[test]
-    fn test_watch_response_encoding() {
-        fluvio_future::subscriber::init_logger();
-        let update = MetadataUpdate {
-            epoch: 2,
-            changes: vec![],
-            all: vec![],
-        };
-        let watch_response: WatchResponse<TopicSpec> = WatchResponse::new(update);
-
-        let mut src = vec![];
-        watch_response
-            .encode(&mut src, ObjectApiWatchRequest::API_KEY as i16)
-            .expect("encoding");
-        //watch_response.encode(&mut src, 0).expect("encoding");
-        println!("output: {src:#?}");
-        let dec = WatchResponse::<TopicSpec>::decode_from(
-            &mut Cursor::new(&src),
-            ObjectApiWatchRequest::API_KEY as i16,
-        )
-        .expect("decode");
-        assert_eq!(dec.inner().epoch, 2);
-    }
-
-    #[test]
-    fn test_obj_watch_response_encode_decoding() {
-        fluvio_future::subscriber::init_logger();
-
-        let res = create_res();
-
-        let mut header = RequestHeader::new(ObjectApiWatchRequest::API_KEY);
-        header.set_client_id("test");
-        header.set_correlation_id(11);
-        let res_msg = ResponseMessage::from_header(&header, res);
-
-        let mut src = vec![];
-        res_msg
-            .encode(&mut src, ObjectApiWatchRequest::API_KEY as i16)
-            .expect("encoding");
-
-        println!("output: {src:#?}");
-
-        assert_eq!(
-            src.len(),
-            res_msg.write_size(ObjectApiWatchRequest::API_KEY as i16)
-        );
-
-        let dec_msg: ResponseMessage<ObjectApiWatchResponse> = ResponseMessage::decode_from(
-            &mut Cursor::new(&src),
-            ObjectApiWatchRequest::API_KEY as i16,
-        )
-        .expect("decode");
-        let _ = (dec_msg.response.downcast().expect("downcast")
-            as Option<WatchResponse<TopicSpec>>)
-            .unwrap();
-    }
-
-    #[test]
-    fn test_obj_watch_api_decoding() {
-        fluvio_future::subscriber::init_logger();
-
-        let res = create_res();
-
-        let mut header = RequestHeader::new(ObjectApiWatchRequest::API_KEY);
-        header.set_client_id("test");
-        header.set_correlation_id(11);
-        let res_msg = ResponseMessage::from_header(&header, res);
-
-        let mut src = vec![];
-        res_msg
-            .encode(&mut src, ObjectApiWatchRequest::API_KEY as i16)
-            .expect("encoding");
-
-        println!("output: {src:#?}");
-
-        assert_eq!(
-            src.len(),
-            res_msg.write_size(ObjectApiWatchRequest::API_KEY as i16)
-        );
-
-        let dec_msg: ResponseMessage<ObjectApiWatchResponse> = ResponseMessage::decode_from(
-            &mut Cursor::new(&src),
-            ObjectApiWatchRequest::API_KEY as i16,
-        )
-        .expect("decode");
-        let _ = (dec_msg.response.downcast().expect("downcast")
-            as Option<WatchResponse<TopicSpec>>)
-            .unwrap();
-    }
-
-    #[test]
-    fn test_list_response_encode_decoding() {
-        use fluvio_protocol::api::Request;
-
-        fluvio_future::subscriber::init_logger();
-
-        let list = ListResponse::<CustomSpuSpec>::new(vec![Metadata {
-            name: "test".to_string(),
-            spec: CustomSpuSpec::default(),
-            status: SpuStatus::default(),
-        }]);
-
-        let resp = ObjectApiListResponse::try_encode_from(list, COMMON_VERSION).expect("encode");
-
-        let mut header = RequestHeader::new(ObjectApiListRequest::API_KEY);
-        header.set_client_id("test");
-        header.set_correlation_id(11);
-        let res_msg = ResponseMessage::from_header(&header, resp);
-        let mut src = vec![];
-        res_msg.encode(&mut src, COMMON_VERSION).expect("encoding");
-
-        println!("output: {src:#?}");
-
-        let dec_msg: ResponseMessage<ObjectApiListResponse> =
-            ResponseMessage::decode_from(&mut Cursor::new(&src), COMMON_VERSION).expect("decode");
-
-        let response = (dec_msg.response.downcast().expect("downcast")
-            as Option<ListResponse<CustomSpuSpec>>)
-            .unwrap();
-        assert_eq!(response.inner().len(), 1);
-    }
-}
+mod test;
